Replaced the {dbo_table}'s on the view invoice page with {form_table}'s

+++ manager/pages/ViewInvoicePage.class.php
@@ -39,13 +39,29 @@ class ViewInvoicePage extends SolidStatePage
     // Give the template acess to the invoice DBO
     $this->session['invoice_dbo'] =& $this->get['invoice'];
 
+    $invoiceID = $this->get['invoice']->getID();
+
     // Set the invoice URL field
-    $this->setURLField( "invoice", $this->get['invoice']->getID() );
+    $this->setURLField( "invoice", $invoiceID );
 
     // Set this page's Nav Vars
     $this->setNavVar( "account_id", $this->get['invoice']->getAccountID() );
     $this->setNavVar( "account_name", $this->get['invoice']->getAccountName() );
-    $this->setNavVar( "invoice_id", $this->get['invoice']->getID() );
+    $this->setNavVar( "invoice_id", $invoiceID );
+
+    // Setup the Invoice Item table
+    $iiField = $this->forms['view_invoice_items']->getField( "items" );
+    $iiField->getWidget()->setInvoiceID( $invoiceID );
+    $iiField->getValidator()->setInvoiceID( $invoiceID );
+    
+    // Setup the Payment table
+    $ptField = $this->forms['view_invoice_payments']->getField( "payments" );
+    $ptField->getWidget()->setInvoiceID( $invoiceID );
+    $ptField->getValidator()->setInvoiceID( $invoiceID );
+
+    // Setup the Outstanding Invoices Table
+    $oiField = $this->forms['view_invoice_outstanding_invoices']->getField( "invoices" );
+    $oiField->getWidget()->setPriorToInvoice( $this->get['invoice'] );
   }
 
   /**
@@ -100,12 +116,18 @@ class ViewInvoicePage extends SolidStatePage
 	$this->add_line_item();
 	break;
 
-      case "delete_item":
-	$this->delete_line_item();
+      case "view_invoice_items":
+	if( isset( $this->post['remove'] ) )
+	  {
+	    $this->delete_line_item();
+	  }
 	break;
 
-      case "delete_payment":
-	$this->delete_payment();
+      case "view_invoice_payments":
+	if( isset( $this->post['remove'] ) )
+	  {
+	    $this->delete_payment();
+	  }
 	break;
 
       default:
@@ -121,18 +143,15 @@ class ViewInvoicePage extends SolidStatePage
    */
   function delete_payment()
   {
-    // Verify that a payment was provided
-    if( !isset( $this->get['payment'] ) )
+    // Delete Payments
+    foreach( $this->post['payments'] as $dbo )
       {
-	throw new SWException( "There is no payment to delete!" );
-      }
-
-    // Delete PaymentDBO
-    if( !delete_PaymentDBO( $this->get['payment'] ) )
-      {
-	// Could not delete payment
-	$this->setError( array( "type" => "DB_DELETE_PAYMENT_FAILED" ) );
-	$this->reload();
+	if( !delete_PaymentDBO( $dbo ) )
+	  {
+	    // Could not delete payment
+	    $this->setError( array( "type" => "DB_DELETE_PAYMENT_FAILED" ) );
+	    $this->reload();
+	  }
       }
 
     // Success
@@ -147,24 +166,20 @@ class ViewInvoicePage extends SolidStatePage
    */
   function delete_line_item()
   {
-    // Verify that a line item was provided
-    if( !isset( $this->get['item'] ) )
+    // Delete Invoice Items
+    foreach( $this->post['items'] as $dbo )
       {
-	throw new SWException( "There is no line item to delete!" );
-      }
-    
-    // Delete InvoiceItemDBO
-    if( !delete_InvoiceItemDBO( $this->get['item'] ) )
-      {
-	// Could not delete line item
-	$this->setError( array( "type" => "DB_DELETE_INVOICE_ITEM_FAILED",
-				"args" => array( $item_dbo->getText() ) ) );
-	$this->reload();
+	if( !delete_InvoiceItemDBO( $dbo ) )
+	  {
+	    // Could not delete line item
+	    $this->setError( array( "type" => "DB_DELETE_INVOICE_ITEM_FAILED",
+				    "args" => array( $dbo->getText() ) ) );
+	    $this->reload();
+	  }
       }
 
     // Success
-    $this->setMessage( array( "type" => "INVOICE_ITEM_DELETED",
-			      "args" => array( $this->get['item']->getText() ) ) );
+    $this->setMessage( array( "type" => "INVOICE_ITEM_DELETED" ) );
     $this->reload();
   }
 
@@ -191,16 +206,6 @@ class ViewInvoicePage extends SolidStatePage
 
     // Success
     $this->setMessage( array( "type" => "INVOICE_ITEM_CREATED" ) );
-  }
-
-  /**
-   * Populate Outstanding Invoices Table
-   *
-   * @return array InvoiceDBO's
-   */
-  function populateOutstandingInvoices()
-  {
-    return $this->get['invoice']->getOutstandingInvoices();
   }
 }
